requirements.txt
flask==2.0.1
google-generativeai==0.3.0
transformers==4.30.2
requests==2.31.0